ManipurSDLC Spider Complete and Run.py Running
Completed all three spiders for ManipurSLDC and got run.py to run all three of the spiders at once. Commented out the last part of ChhattisgarhSLDC (the code that was done to run multiple spiders at once) to allow the rest of the spiders to work. Renamed some files as well.